円の面積機能追加
