--- app.py
@@ -1,10 +1,0 @@
-from flask import Flask, render_template
-
-app = Flask(__name__)
-
-@app.route("/")
-def top_page():
-    return render_template("index.html")
-
-if __name__ == "__main__":
-    app.run(debug=True)
