+++ templates/circle_input.html
@@ -1,0 +1,9 @@
+<html>
+    <body>
+        <form action="{{url_for('circle_result')}}" method="get">
+        半径： <input type="text" name="radius"><br>
+        <input type="submit" value="計算">
+        </form>
+    </body>
+
+</html>
