+++ templates/circle_result.html
@@ -1,0 +1,5 @@
+<html>
+    <body>
+        円の面積は{{result}}です。
+    </body>    
+</html>
