+++ templates/index.html
@@ -2,5 +2,6 @@
 <html lang="ja">
 <body>
         ここはトップページです。<br>
+        <a href="{{url_for('circle_input')}}">面積計算(円)</a><br>
 </body>
 </html>
